refactor: Sources' REST API client
The REST API functions that were being used were almost a copy paste of
one another, which made following the code difficult.

The goal of the refactor was to simplify the code and make it easier to
follow and reason about. Also, it introduces a timeout for the requests
by default, which could be the cause for which the Superkey struggles so
much and ends up restarting non-stop.

Another goal of this refactor was to leave everything in an state where
it is easier to reason about how the Superkey worker should be
refactored further, like introducing proper dependency injection and
untangling the resource making from the "ForgeApplication" struct.
However, that's for another PR.

RHCLOUD-35843